fix: script
